+++ src/main/java/com/idea5/four_cut_photos_map/security/WebConfig.java
@@ -1,13 +1,19 @@
 package com.idea5.four_cut_photos_map.security;
 
+import com.idea5.four_cut_photos_map.security.jwt.interceptor.AuthenticationInterceptor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @EnableWebMvc
+@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
+    private final AuthenticationInterceptor authenticationInterceptor;
+
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
@@ -15,5 +21,13 @@ public class WebConfig implements WebMvcConfigurer {
                 .allowedMethods("*")
                 .allowCredentials(true)
                 .maxAge(3000);
+    }
+
+    // 인터셉터 등록
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(authenticationInterceptor)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/auth/login/kakao", "/shops/**", "/auth/token"); // 인가작업을 제외할 API 경로 설정
     }
 }
